# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Socal Mediation Center API",
        "description": "REST APIs for Socal mediation and legal consultation booking flows.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Application API"
        }
    ],
    "paths": {
...
        "/v1/consultations/draft": {
            "post": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Create or save a draft consultation",
                "description": "Creates the draft after consultation type selection. Optional fields can be sent from the details form to preserve partially completed data while the status remains draft.",
                "operationId": "f24f08a3bb792234caea075eae9f8dbf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "consultation_type_id"
                                ],
                                "properties": {
                                    "consultation_type_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "legal_service_name": {
                                        "type": "string",
                                        "example": "Business, Payment & Contract Disputes",
                                        "nullable": true
                                    },
                                    "consultation_mode": {
                                        "type": "string",
                                        "example": "online",
                                        "nullable": true,
                                        "enum": [
                                            "online",
                                            "offline",
                                            "phone"
                                        ]
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Briefly describe the legal guidance or assistance needed.",
                                        "nullable": true
                                    },
                                    "referral_source": {
                                        "type": "string",
                                        "example": "Google",
                                        "nullable": true
                                    },
                                    "referral_source_others": {
                                        "description": "Used only when referral_source is Other Referral.",
                                        "type": "string",
                                        "example": "Friend at local business group",
                                        "nullable": true
                                    },
                                    "primary_client": {
                                        "$ref": "#/components/schemas/ParticipantPayload"
                                    },
                                    "participants": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/ParticipantPayload"
                                        },
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Draft consultation created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Draft consultation created."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Consultation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/consultations/{consultation}": {
            "get": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Get a consultation by UUID",
                "operationId": "7afb8bdee805c3e8498845e4f0f4c383",
                "parameters": [
                    {
                        "name": "consultation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consultation detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Consultation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Consultation not found"
                    }
                }
            }
        },
...
        "/v1/consultations/{consultation}/reschedule": {
            "post": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Reschedule a confirmed consultation booking",
                "description": "Updates the selected date/time after checking availability. Online bookings get a regenerated Zoom meeting link, and Outlook sync recreates the calendar event so the old event is removed.",
                "operationId": "e356139d41c05f69320d78cab77ac6c5",
                "parameters": [
                    {
                        "name": "consultation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "starts_at"
                                ],
                                "properties": {
                                    "starts_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-08-14T09:00:00-07:00"
                                    },
                                    "timezone": {
                                        "type": "string",
                                        "example": "America/Los_Angeles",
                                        "nullable": true
                                    },
                                    "professional_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Consultation rescheduled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Consultation rescheduled."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Consultation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unavailable slot, inactive booking, or validation failed"
                    }
                }
            }
        },
        "/v1/consultations/{consultation}/reschedule-status": {
            "get": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Check whether a consultation can continue through reschedule flow",
                "description": "Returns only pending or completed. Completed consultations should not be rescheduled.",
                "operationId": "38b3e5f109642ef5c9204b044fb6228d",
                "parameters": [
                    {
                        "name": "consultation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reschedule status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending",
                                                    "enum": [
                                                        "pending",
                                                        "completed"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Consultation not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ParticipantPayload": {
                "required": [
                    "first_name"
                ],
                "properties": {
                    "first_name": {
                        "type": "string",
                        "example": "Jonathan"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Miller",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jonathan@example.com",
                        "nullable": true
                    },
                    "phone_country": {
                        "type": "string",
                        "example": "+1",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "example": "(495) 060-0000",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}